do not nag about nonexisting _main_test.c

+++ all.test.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 #C_FILES=$(find . -maxdepth 1 -name \*.c|sed '/_test\.c/d;/tap\./d;/logcat\.c/d'| tr '\n' ' ')
 C_FLAGS="$C_FLAGS -ggdb -Wall -Wextra -Werror"
-rm _main_test.c
+[ -f _main_test.c ] && rm _main_test.c
 TEST_FILES=$(find . -maxdepth 1 -name \*_test.c)
 for testfile in $TEST_FILES; do
 	TESTFUN=$(basename $testfile .c)
